Refactor PeepActionSpriteType to use strong enum

+++ src/openrct2/paint/sprite/Paint.Peep.cpp
@@ -80,7 +80,8 @@ void peep_paint(paint_session* session, const Peep* peep, int32_t imageDirection
 
     // In the following 4 calls to sub_98197C/sub_98199C, we add 5 (instead of 3) to the
     //  bound_box_offset_z to make sure peeps are drawn on top of railways
-    uint32_t baseImageId = (imageDirection >> 3) + sprite.sprite_animation[spriteType].base_image + imageOffset * 4;
+    uint32_t baseImageId = (imageDirection >> 3) + sprite.sprite_animation[static_cast<uint8_t>(spriteType)].base_image
+        + imageOffset * 4;
     uint32_t imageId = baseImageId | peep->TshirtColour << 19 | peep->TrousersColour << 24 | IMAGE_TYPE_REMAP
         | IMAGE_TYPE_REMAP_2_PLUS;
     sub_98197C(session, imageId, 0, 0, 1, 1, 11, peep->z, 0, 0, peep->z + 5);
